+++ lib/lifer/builder.rb
@@ -101,4 +101,5 @@ end
 
 require_relative "builder/rss"
 require_relative "builder/html"
+require_relative "builder/json_feed"
 require_relative "builder/txt"